+++ CHANGELOG.md
@@ -9,6 +9,9 @@
 - Spiffed up the role menu commands, and added support for emojis on role labels.
 - Base capture notifications now show the amount of outfit resource earned.
 
+**Deprecations**
+- Removed the 'most used weapon' field from the `character` command. It rarely displayed a relevant weapon.
+
 **Fixes**
 - Fixed Oshur status being calculated incorrectly.
 - Fixed the character name autocomplete on the `character` command failing when not using all-lowercase letters.